# apps/web/src/lib/service-clients/service-auth/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "authentication_service",
    "description": "",
    "termsOfService": "https://macro.com/terms",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/permissions": {
      "get": {
        "tags": ["permissions::get_permissions"],
        "summary": "Gets all permissions",
        "operationId": "get_permissions",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Permission"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/permissions/me": {
      "get": {
        "tags": ["permissions::get_user_permissions"],
        "summary": "Gets a list of the users permissions",
        "operationId": "get_user_permissions",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/user/me": {
      "get": {
        "tags": ["user::get_user_info"],
        "summary": "Gets the calling user's info",
        "operationId": "get_user_info",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserInfo"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["user::delete_user"],
        "summary": "Deletes the user who calls this endpoint",
        "operationId": "delete_user",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/user/tutorial": {
      "patch": {
        "tags": ["user::patch_tutorial"],
        "summary": "Updates the user's tutorialComplete flag.",
        "operationId": "patch_user_tutorial",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchUserTutorialRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "EmptyResponse": {
        "type": "object",
        "description": "Empty response is required due to custom fetch forcing `response.json()`"
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "A plain old json error response for use with axum.\nyup, thats it.",
        "required": ["message"],
        "properties": {
          "message": {
            "type": "string",
            "description": "Message to explain failure"
          }
        }
      },
...
      "GenericSuccessResponse": {
        "type": "object",
        "required": ["success"],
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indicates if the request was successful"
          }
        }
      },
...
      "GetUserInfo": {
        "type": "object",
        "required": ["user_id", "permissions"],
        "properties": {
          "organization_id": {
            "type": ["integer", "null"],
            "format": "int32",
            "description": "The user's organization id if there is one associated with the user"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The user's permissions"
          },
          "user_id": {
            "type": "string",
            "description": "The user id"
          }
        }
      },
...
      "PatchUserTutorialRequest": {
        "type": "object",
        "required": ["tutorialComplete"],
        "properties": {
          "tutorialComplete": {
            "type": "boolean",
            "description": "If the user has completed the tutorial"
          }
        }
      },
      "Permission": {
        "type": "object",
        "required": ["id", "description"],
        "properties": {
          "description": {
            "type": "string",
            "description": "The description of the permission"
          },
          "id": {
            "type": "string",
            "description": "The permission"
          }
        }
      },
...
    }
  },
...
}